+++ src/Game/Text/el/ScreenText.el.sjson.lua
@@ -32,14 +32,6 @@ local newData = {
 	-- #endregion
 
 	-- #region Codex/Inventory
-	{
-		Id = "ModsNikkelMHadesBiomesCodexEntryTitleText",
-		InheritFrom = "WeaponPointsRare",
-	},
-	{
-		Id = "InventoryScreen_ModesNikkelMHadesBiomesTab",
-		InheritFrom = "ModsNikkelMHadesBiomesCodexEntryTitleText",
-	},
 	-- {
 	-- 	Id = "ModsNikkelMHadesBiomes_MarketScreen_Resources",
 	-- 	DisplayName = "Barter Exchange",